# Back-end/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Gym Management API",
    "description": "Sistema de control de acceso y gestión de usuarios para el gimnasio UES",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "cookie",
      "name": "token",
      "description": "Token de sesión almacenado en cookies"
    }
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Iniciar sesión",
        "description": "Verifica credenciales y genera una HTTP-Only Cookie con el token JWT.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Credenciales de acceso",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "admin@gym.com"
                },
                "password": {
                  "type": "string",
                  "example": "12345"
                }
              },
              "required": [
                "email",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login exitoso",
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "$ref": "#/definitions/AdminResponse"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Email o contraseña incorrectos."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Obtener perfil actual",
        "description": "Retorna la información del administrador autenticado basándose en el token de la cookie.",
        "responses": {
          "200": {
            "description": "Información del usuario actual.",
            "schema": {
              "$ref": "#/definitions/AdminResponse"
            }
          },
          "401": {
            "description": "No autorizado / Token inválido."
          }
        }
      }
    },
...
  },
...
}